Refactor CSR elementer
